# Conflicts:
#	Assets/Sprites/UI/Barra/Mana/EvilManaGlow.psd.meta
#	Assets/Sprites/UI/Barra/Mana/Sparkle.png.meta
#	Assets/Sprites/UI/Barra/Mana/evilFrameBackground.psd.meta
#	Assets/Sprites/UI/Barra/Mana/evilFrameBorders.psd.meta
#	Assets/Sprites/UI/Barra/Mana/evilFrameFill.png.meta
#	Assets/Sprites/UI/Barra/Mana/evilFramePotionIcon.png.meta
#	Assets/Sprites/UI/Mana/EvilManaGlow.psd
#	Assets/Sprites/UI/Mana/EvilManaGlow.psd.meta
#	Assets/Sprites/UI/Mana/evilFrameBackground.psd
#	Assets/Sprites/UI/Mana/evilFrameBackground.psd.meta
#	Assets/Sprites/UI/Mana/evilFrameBorders.psd
#	Assets/Sprites/UI/Mana/evilFrameBorders.psd.meta
#	Assets/Sprites/UI/Mana/evilFrameFill.psd
#	Assets/Sprites/UI/Mana/evilFrameFill.psd.meta
#	Assets/Sprites/UI/Mana/evilFrameIcon.psd
#	Assets/Sprites/UI/Mana/evilFrameIcon.psd.meta

+++ Assets/Scripts/General/IUsable.cs
@@ -3,7 +3,7 @@
     // Called by Player
     bool CanUse();
     // Called by Player
-    int GetUsageCost();
+    float GetUsageCost();
     // Called by Player
     bool Activate(Player player);
     // Called by Player
